// src/app/models/software-catalog.d.ts
export interface components {
    schemas: {
// ...
        Characteristic: {
            /** @description Unique identifier of the characteristic */
            id?: string;
            /** @description Name of the characteristic */
            name: string;
            /** @description Data type of the value of the characteristic */
            valueType?: string;
            characteristicRelationship?: components["schemas"]["CharacteristicRelationship"][];
            value: components["schemas"]["Any"];
            /** @description When sub-classing, this defines the super-class */
            "@baseType"?: string;
            /**
             * Format: uri
             * @description A URI to a JSON-Schema file that defines additional attributes and relationships
             */
            "@schemaLocation"?: string;
            /** @description When sub-classing, this defines the sub-class Extensible name */
            "@type"?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}